Fixed a bug where the MQTT connection would not reconnect when the phone dropped a data connection (wifi and cellular)
It turns out the initial connection error and the dropped connection errors are identical, so it will require more code to know if this was a user initiated connection or a reconnect based on a dropped connection (or an on boot connection). For now we will just not set the status to disconnected.